[#1087] Simplified logic of Buttons component and fixed naming bug in the Message component.
