+++ frontend/src/common/confirm-dialog/Message.tsx
@@ -1,13 +1,13 @@
 import * as React from 'react';
 
-interface ConfirmMessageProps {
+interface MessageProps {
   message: React.ReactNode;
 }
 
-const ConfirmMessage: React.StatelessComponent<ConfirmMessageProps> = ({ message }) => (
+const Message: React.StatelessComponent<MessageProps> = ({ message }) => (
   <div className='ConfirmDialog-message'>
     {message}
   </div>
 );
 
-export default ConfirmMessage;
+export default Message;
